FIX: Prevent sardalign dependencies from being installed
When installing sardalign, the fairseq dependency is a very old and little updated repo, which requires an outdated version of OmegaConf and consequently a downgrade of the pip installer package.

This causes a bug when running tune run <recipe> <config> because the API of OmegaConf has been updated since the version used by fairseq.

Note: We cannot specify the --no-deps flag inside the requirements.txt file. See:
- https://stackoverflow.com/a/31798428/6619692
- Stalled PR that would provide this functionality: https://github.com/pypa/pip/pull/10837